改进模块写法
