+++ 302-backbone-quiz-app/index.html
@@ -18,7 +18,7 @@
 <!--页面模板-->
 <script type="text/template" id="card_template">
     <a href=<%= hrefA %> class="answer_a btn btn-primary btn-block"><%= answerA %></a>
-    <a href=<%= hrefB %> class="answer_a btn btn-default btn-block"><%= answerB %></a>
+    <a href=<%= hrefB %> class="answer_b btn btn-default btn-block"><%= answerB %></a>
 </script>
 
 <script src="http://cdn.staticfile.org/jquery/2.1.1-rc2/jquery.min.js"></script>
@@ -26,6 +26,6 @@
 <script src="http://cdn.staticfile.org/json2/20140204/json2.min.js"></script>
 <script src="http://cdn.staticfile.org/underscore.js/1.7.0/underscore-min.js"></script>
 <script src="http://cdn.staticfile.org/backbone.js/1.1.2/backbone-min.js"></script>
-<script src="dist/js/main.min.js"></script>
+<script src="dist/js/main.js"></script>
 </body>
 </html>
