+++ 302-backbone-quiz-app/src/js/model.js
@@ -1,0 +1,10 @@
+let Card = Backbone.Model.extend({
+    defaults: {
+        imageURL: '',
+        question: '',
+        answerA : '',
+        answerB : ''
+    }
+});
+
+export default Card;
